+++ src/model2/factories/index.ts
@@ -1,4 +1,5 @@
 export * from "./account_factory";
 export * from "./account_flags_factory";
 export * from "./account_thresholds_factory";
+export * from "./account_values_factory";
 export * from "./signer_factory";
